+++ optaweb-vehicle-routing-backend/src/main/java/org/optaweb/vehiclerouting/plugin/planner/domain/DistanceMap.java
@@ -23,9 +23,10 @@ package org.optaweb.vehiclerouting.plugin.planner.domain;
 public interface DistanceMap {
 
     /**
-     * Get distance from a reference location to the given location.
+     * Get distance from a reference location to the given location. The actual physical quantity (distance or time)
+     * and its units depend on the configuration of the routing engine and is not important for optimization.
      * @param location location the distance of which will be returned
      * @return location's distance
      */
-    double distanceTo(PlanningLocation location);
+    long distanceTo(PlanningLocation location);
 }
